+++ src/main/java/ru/yandex/practicum/filmorate/storage/MpaStorage.java
@@ -5,9 +5,10 @@ import ru.yandex.practicum.filmorate.model.Mpa;
 import java.util.Collection;
 
 public interface MpaStorage {
-    Mpa getMpaById(Integer mpaId);
+    Mpa getMpaById(Long mpaId);
 
     Collection<Mpa> findAll();
 
-    boolean isMpaExists(Integer mpaId);
+    boolean isMpaExists(Long mpaId);
+
 }
